tyrquake: fix fp32 cl.time mod precision in R_SetSkyFrame
R_SetSkyFrame computes the wrapped sky animation time as

    skytime = cl.time - ((int)(cl.time / temp) * temp);

with 'temp' declared 'float', so the intermediate (int) * temp runs
in fp32 and the chain narrows before the subtraction. Same antipattern
as V_CalcBob fixed earlier in the audit: ulp at cl.time ~ 1e6 (~12 d
of continuous single-level engine time) is ~0.06 in float space,
which makes the sky-scroll phase quantise visibly.

Promote 'temp' to double so the mod-and-multiply chain stays in
double; cast the bounded final result (< temp, which is a small
integer in practice -- SKYSIZE * s1 * s2 with SKYSIZE=128 and small
sky-speed factors) back to fp32 at the skytime assignment site
where the cast is exact.

+++ common/r_sky.c
@@ -57,7 +57,13 @@ void
 R_SetSkyFrame(void)
 {
     int g, s1, s2;
-    float temp;
+    /* 'temp' is the modulus we wrap cl.time against. Keep it (and the
+     * mod-then-multiply chain) in double so cl.time-derived inputs
+     * keep full precision regardless of how long the engine has been
+     * running; the final result is bounded below temp, so the cast
+     * to fp32 for skytime storage is exact. Same antipattern as
+     * V_CalcBob's bob-cycle math fixed earlier in the series. */
+    double temp;
 
     skyspeed = iskyspeed;
     skyspeed2 = iskyspeed2;
@@ -67,5 +73,5 @@ R_SetSkyFrame(void)
     s2 = iskyspeed2 / g;
     temp = SKYSIZE * s1 * s2;
 
-    skytime = cl.time - ((int)(cl.time / temp) * temp);
+    skytime = (float)(cl.time - ((int)(cl.time / temp) * temp));
 }
